refactor: relocate MatchingQueryService to application package
